feature/ATC-36 - Renames AreaModel to AirspaceModel * Updates references * Moves AirspaceModel to airport forlder
